[feature/staking] update staking screen
